package/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "pm-os"
dynamic = ["version"]
description = "PM-OS: AI-powered Product Management Operating System"
readme = "README.md"
license = "MIT"
requires-python = ">=3.10"
authors = [
    {name = "PM-OS Team"}
]
keywords = ["pm", "product-management", "ai", "claude", "llm"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Software Development :: Libraries :: Python Modules",
]

dependencies = [
    "pyyaml>=6.0",
    "python-dotenv>=1.0.0",
    "requests>=2.28.0",
    "click>=8.0.0",
    "rich>=13.0.0",
]

[project.optional-dependencies]
google = [
    "google-api-python-client>=2.0.0",
    "google-auth>=2.0.0",
    "google-auth-oauthlib>=1.0.0",
]
slack = ["slack-sdk>=3.0.0"]
jira = ["jira>=3.5.0"]
github = ["PyGithub>=1.59.0"]
confluence = ["atlassian-python-api>=3.0.0"]
ai = [
    "anthropic>=0.18.0",
    "google-generativeai>=0.8.0",
]
bedrock = [
    "boto3>=1.28.0",
]
all = [
    "pm-os[google,slack,jira,github,confluence,ai,bedrock]"
]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
]

[project.scripts]
pm-os = "pm_os.cli:main"
pmos = "pm_os.cli:main"

[project.urls]
Homepage = "https://github.com/feamando/pmos"
Documentation = "https://pm-os.dev/docs"
Repository = "https://github.com/feamando/pmos"

[tool.hatch.version]
path = "VERSION"
pattern = "(?P<version>.+)"

[tool.hatch.build.targets.wheel]
packages = ["src/pm_os"]

[tool.hatch.build.targets.sdist]
include = [
    "src/pm_os/**/*.py",
    "src/pm_os/**/*.yaml",
    "src/pm_os/**/*.json",
    "src/pm_os/**/*.md",
    "src/pm_os/templates/*",
    "VERSION",
    "README.md",
    "LICENSE",
]

[tool.black]
line-length = 100
target-version = ["py310", "py311", "py312"]

[tool.ruff]
line-length = 100
select = ["E", "F", "W", "I", "N", "UP", "B", "C4"]
ignore = ["E501"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
